requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
boto3==1.28.10
botocore==1.31.10
Django==4.2.3
django-autoslug==1.9.9
django-cities-light==3.9.2
django-ckeditor==6.6.1
django-js-asset==2.1.0
djangorestframework==3.14.0
drf-yasg==1.21.7
inflection==0.5.1
jmespath==1.0.1
mysqlclient==2.2.0
packaging==23.1
progressbar2==4.2.0
python-dateutil==2.8.2
python-utils==3.7.0
pytz==2023.3
PyYAML==6.0.1
s3transfer==0.6.1
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
typing_extensions==4.7.1
Unidecode==1.3.6
uritemplate==4.1.1
urllib3==1.26.16
